Possible fix for unaligned overTime client shared memory
`counters->overTime` may not be a multiple of `pagesize`, so the
original code may have caused the overTime client memory size to stop
being a multiple of the page size. We have seen issues with shared
memory that is not a multiple of the page size in the past, and this may
be causing the current issues in v4.2 and v4.2.1.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>